roby: don't unconditionally propagate use_frames selection in the task hierarchy
The old behaviour was to propagate frame selections in the task hierarchy
unconditionally. A (arguably better) behaviour implemented in this commit
is to stop propagating at each level where something has been defined.

For instance, with

use Cmp::PoseEstimation => Cmp::PoseEstimation.
  use_frames("world" => "world", "body" => "body")
use Cmp::Odometry => Cmp::Odometry.
  use_frames("odometry" => "odometry", "body" => "body")

one would get an error with the old behaviour, as the "world" selection
gets propagated to the Cmp::Odometry that is part of the pose estimation.

In the new behaviour, the world selection is stopped at the odometry
(as Cmp::Odometry has explicit frame selections) and the frame assignation
goes well.